move new bloom filter implementation over
